requirements.txt
python-telegram-bot
telethon
transformers
torch
numpy 
matplotlib.pyplot
